# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: Portal
    description: Client portal endpoints (authenticated clients see only their own documents).
  - name: Admin
    description: Admin management endpoints (upload and assign documents to clients).
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /portal/me:
    get:
      operationId: getMe
      tags: [Portal]
      summary: Get the current authenticated user's profile
      responses:
        "200":
          description: Current user profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /portal/documents:
    get:
      operationId: listMyDocuments
      tags: [Portal]
      summary: List documents owned by the current user
      responses:
        "200":
          description: The current user's documents.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentItem"
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /portal/documents/{id}/content:
    get:
      operationId: getMyDocumentContent
      tags: [Portal]
      summary: Stream the content of a document owned by the current user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document content streamed with its stored Content-Type.
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Document not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [Admin]
      summary: Request a presigned URL for file upload
      description: |
        Returns a presigned GCS URL for direct upload. The client sends JSON
        metadata here, then uploads the file directly to the returned URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/documents:
    get:
      operationId: listAdminDocuments
      tags: [Admin]
      summary: List all documents with owner information
      responses:
        "200":
          description: All documents.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminDocumentItem"
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      operationId: createDocument
      tags: [Admin]
      summary: Create a document and assign it to a client by email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Document created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDocumentItem"
        "400":
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/documents/{id}:
    delete:
      operationId: deleteDocument
      tags: [Admin]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Document deleted.
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Document not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/clients:
    get:
      operationId: listClients
      tags: [Admin]
      summary: List all client users
      responses:
        "200":
          description: All client users.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClientItem"
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    UserProfile:
      type: object
      required: [id, email, role]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: ["string", "null"]
        role:
          type: string
    DocumentItem:
      type: object
      required: [id, title, contentType, size, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        contentType:
          type: string
        size:
          type: integer
        createdAt:
          type: string
          format: date-time
    AdminDocumentItem:
      type: object
      required: [id, title, contentType, size, createdAt, ownerEmail]
      properties:
        id:
          type: integer
        title:
          type: string
        contentType:
          type: string
        size:
          type: integer
        createdAt:
          type: string
          format: date-time
        ownerEmail:
          type: string
        ownerName:
          type: ["string", "null"]
    ClientItem:
      type: object
      required: [id, email, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    DocumentInput:
      type: object
      required: [title, ownerEmail, objectPath, contentType, size]
      properties:
        title:
          type: string
          minLength: 1
        ownerEmail:
          type: string
          minLength: 1
        objectPath:
          type: string
          minLength: 1
        contentType:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 0
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"