# lib/api-zod/src/generated/types/adminDocumentItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminDocumentItem {
  id: number;
  title: string;
  contentType: string;
  size: number;
  createdAt: Date;
  ownerEmail: string;
  /** @nullable */
  ownerName?: string | null;
}