# lib/api-zod/src/generated/types/clientItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ClientItem {
  id: number;
  email: string;
  /** @nullable */
  name?: string | null;
  createdAt: Date;
}