# lib/api-zod/src/generated/types/documentInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DocumentInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  ownerEmail: string;
  /** @minLength 1 */
  objectPath: string;
  /** @minLength 1 */
  contentType: string;
  /** @minimum 0 */
  size: number;
}