# lib/api-zod/src/generated/types/documentItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DocumentItem {
  id: number;
  title: string;
  contentType: string;
  size: number;
  createdAt: Date;
}